+++ packages/connection/src/common/index.ts
@@ -1,7 +1,7 @@
 export * from './message';
 export * from './rpc-service/proxy';
 export * from './rpc/multiplexer';
-export * from './ext-rpc-protocol';
+export * from './rpcProtocol';
 export * from './utils';
 export * from './ws-channel';
 export * from './connect';
