+++ packages/connection/src/common/rpcProtocol.ts
